ai-engine/requirements.txt
pandas
numpy
scikit-learn
xgboost
pymongo
python-dotenv
joblib
